# openapi.yaml
openapi: 3.0.0
info:
  title: Bank Rates API
  description: API for accessing information about banks.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /rates:
    post:
      summary: Retrieve bank rates
      description: Retrieve bank rates for specified bank IDs, currency ID, and date.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bankIds:
                  type: array
                  items:
                    type: integer
                  description: List of bank IDs
                currencyId:
                  type: integer
                  description: Currency ID
                date:
                  type: string
                  format: date
                  description: Date in YYYY-MM-DD format. If not provided, current date is used.
      responses:
        '200':
          description: Bank rates retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankRate'
        '500':
          description: Internal server error
  /banks:
    get:
      summary: Retrieve banks
      description: |
        Retrieve a list of banks.
      responses:
        '200':
          description: A list of banks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bank'
        '500':
          description: Internal server error
    post:
      summary: Create a new bank
      description: |
        Create a new bank entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBank'
      responses:
        '201':
          description: Bank created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bank'
        '500':
          description: Internal server error
    delete:
      summary: Delete a bank by ID
      description: |
        Delete a bank by its ID.
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the bank to delete
          schema:
            type: integer
      responses:
        '204':
          description: Bank deleted successfully
        '404':
          description: Bank not found
        '500':
          description: Internal server error
  /currencies:
    get:
      summary: Retrieve Currencies
      description: |
        Retrieve a list of Currencies.
      responses:
        '200':
          description: A list of Currencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Currency'
        '500':
          description: Internal server error
    post:
      summary: Create a new currency
      description: |
        Create a new currency entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCurrency'
      responses:
        '201':
          description: Currency created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '500':
          description: Internal server error
components:
  schemas:
    Bank:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The ID of the bank.
        name:
          type: string
          description: The name of the bank.
    Currency:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The ID of the currency.
        name:
          type: string
          description: The name of the currency.
        symbol:
          type: string
          description: The symbol of the currency.
    NewBank:
      type: object
      properties:
        name:
          type: string
          description: The name of the new bank.
    NewCurrency:
      type: object
      properties:
        name:
          type: string
          description: The name of the new currency.
        symbol:
          type: string
          description: The symbol of the new currency.
    BankRate:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The ID of the bank rate.
        rate:
          type: number
          description: The rate of the bank.
        date:
          type: string
          format: date
          description: The date of the bank rate.
        bank:
          $ref: '#/components/schemas/Bank'
        currency:
          $ref: '#/components/schemas/Currency'